dir refactor
