+++ hobj/data_loaders/images/highvar.py
@@ -1,5 +1,5 @@
 import mref.media_references
-from hobj.images.template import Imageset
+from hobj.data_loaders.images.template import Imageset
 import pydantic
 
 from typing import Dict, List
