+++ hobj/data_loaders/images/warmup.py
@@ -1,6 +1,6 @@
 import pydantic
 
-from hobj.images.template import Imageset
+from hobj.data_loaders.images.template import Imageset
