--- hobj/images/__init__.py
@@ -1,4 +1,0 @@
-from hobj.images.highvar import MutatorHighVarImageset
-from hobj.images.oneshot import MutatorOneShotImageset
-from hobj.images.warmup import MutatorWarmupImageset
-from hobj.images.probe import ProbeImageset
